+++ routes/web.php
@@ -4,8 +4,8 @@ use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Auth;
 
 Route::get('/', function () {
-    $categories = \App\Models\Category::all()->sortBy('category');
-    return view('categories.index')->with('categories', $categories);
+    $items = \App\Models\Item::all()->sortBy('title');
+    return view('items.index')->with('items', $items);
 });
 
 Auth::routes();
@@ -13,4 +13,6 @@ Auth::routes();
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
 Route::resource('categories', App\Http\Controllers\CategoryController::class);
-Route::get('/categories/delete/{id}', [App\Http\Controllers\CategoryController::class, 'confirmDelete']) -> name('categories.confirmDelete');
+
+Route::resource('items', App\Http\Controllers\ItemController::class);
+Route::get('/items/delete/{id}', [App\Http\Controllers\ItemController::class, 'confirmDelete']) -> name('items.confirmDelete');