+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Changelog
 
+## [0.3] - 2024-12-05
+### Added
+- Navbar links to Items and Categories pages for easy access
+- Form to create item
+- Form to edit item (unfinished)
+- Items table
+- Master list of items
+
+### Changed
+- Color scheme to dark
+- Navbar title links to list of items
+
 ## [0.2] - 2024-12-04
 ### Added
 - Form to create category
-- Edit catagory functionality
+- Edit category functionality
 - Master list of categories
 - Categories Table